Allocations: move Budgets to own component
The Overview component was getting quite large, but most of it was only
there to support the list of Budgets. This shows that more clearly, and
keeps Overview focused.
